# specs/004-secure-env-chatbot/contracts/api-spec.yaml
openapi: 3.0.0
info:
  title: Secure AI Chatbot API
  description: API for secure AI chatbot with RAG functionality for textbook content
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server

paths:
  /api/session/create:
    post:
      summary: Create a new chat session
      description: Creates a new chat session for a user interaction
      responses:
        '200':
          description: Session created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                    description: Unique identifier for the session
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/chat:
    post:
      summary: Send a message to the chatbot
      description: Send a user query to the AI chatbot with session context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                  description: Unique identifier for the session
                message:
                  type: string
                  description: The user's question or message
                text_content:
                  type: string
                  description: The text content to analyze (optional, can be provided by frontend)
              required:
                - session_id
                - message
      responses:
        '200':
          description: Response generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                    description: The AI-generated answer
        '400':
          description: Invalid request (missing session_id or message)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details